requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
PyQt6==6.6.1
PyQt6-Qt6==6.6.3
PyQt6-sip==13.6.0
pyqtgraph==0.13.4
scipy==1.13.0
